added the ability to view a specific page

+++ app/src/main/java/hexlet/code/util/NamedRoutes.java
@@ -9,7 +9,7 @@ public class NamedRoutes {
         return "/urls";
     }
 
-    public static String idUrlPath(Long id) {
+    public static String idUrlPath(int id) {
         return idUrlPath(String.valueOf(id));
     }
